Fix: ui와 container 분리
